Download URL configurable via pillar
I ran into this problem where the formula can't download and install
Consul:

https://github.com/saltstack/salt/issues/38730

As a work around, I'm making the download URL configurable. Not sure if
many other people will find this useful, so if you don't want to merge
it that's fine. For me it's only a temporary fix until I can upgrade
Salt.